+++ src/commands/mod.rs
@@ -19,6 +19,7 @@ pub mod review;
 pub mod send;
 pub mod site;
 pub mod stack;
+pub mod status;
 pub mod task;
 pub mod task_publish;
 pub mod task_run;